refactor : ExceptionCode 위치 변경
