+++ src/main/java/com/example/tily/til/TilService.java
@@ -1,6 +1,6 @@
 package com.example.tily.til;
 
-import com.example.tily._core.errors.ExceptionCode;
+import com.example.tily._core.errors.exception.ExceptionCode;
 import com.example.tily._core.errors.exception.CustomException;
 
 import com.example.tily.comment.Comment;
